sperate drawing and add them to class item
