+++ lib/item.dart
@@ -1,0 +1,25 @@
+void controller() {
+  print("   ____");
+  print(r"  /0__0\");
+}
+
+void tv() {
+  print(" ______");
+  print("||    ||");
+  print("||    ||");
+  print(" ______");
+  print(r"   /\");
+}
+
+void door() {
+  print("\n___");
+  print("[ |");
+  print("---");
+}
+
+void walking() {
+  print("  o ");
+  print(" /");
+  print(" /|");
+  print("${"*" * 14}\n");
+}
